It looks like we're going to have to switch gears...
The range of XMLFragments that I was trying to code up would have worked
with elements, but once you add in stuff like DOCTYPE, it falls flat on
its face, because you then have multiple types of recursion going on
rather than just one, meaning that a stack of state would need to be
maintained, and the code that iterates over the range would have to
maintain a lot of that same state to process the range, making it
incredibly unwieldy, and it would not be able to avoid allocations as
originally envisioned, because it would need to maintain that stack of
state. So, it seemed best to give up on that idea and just directly code
the lowest level that anyone would actually want to operate at with
whole entities being presented at once and in a tree structure rather
than a range - albeit with the idea of passing through it once in a
manner similar to a range rather than actually building a DOM - though
we'll provide a way to build a DOM on top of that too.