Improvements to the Entity/EntityParser API.
EntityParser now looks a lot like a range, but actually going to a range
API doesn't really make sense due to efficiency concerns with making
front not be transient and the efficiency gains that can be had by
having multiple ways to "pop the front" rather than simply going to the
next entity (e.g. skip everything between a start and end tag rather
than iterating to the first entity in its contents).