Remove pos property from EntityParser.
It doesn't really make sense to be exposing the current position of the
parser as part of the API, since that's really an implementaton detail.
Exposing the position of the start of the current entity would make more
sense, but it would add more overhead (though possibly not enough to
matter), it would be yet one more thing that would have to be tested,
and I just can't come up with a good enough use case for it to have it.
We can add it later if we decide that it's really worth having.