Move helper types into EntityCursor.
Doing this makes AttDefInfo be able to work sanely. It needs to be able
to parse an enumeration declaration, and that was going to require
instantiating AttDefInfo with more than just the range type. It was also
getting a bit sticky as to whether the various helper types should be
instantiated with the original range type or SliceOfR. Moving them all
into EntityCursor and putting them next to the properties that they go
with solves all of those problems as well as putting all of the
documentation for them in one place.