The last of the DOCTYPE parsing should now be in place.
Of course, the testing on a lot of it is nonexistent right now, but in
theory, all of the core parsing needed for XML is now in place. Some of
the helper functions on EntityCursor need to be implemented, some free
helper functions using EntityCursor or operating on text returned by one
of its properties are bound to be needed, and lots of tests need to be
added, but in theory, we now have a fully functional STaX parser.