Add more advanced parsing for <!ELEMENT ...> tags.
We still aren't fully parsing the "children" version of the contentspec,
but it's essentially a tree, making it less than straightforward to do
something like just slap a range over it. We'll figure out a way to
deal with it in a more friendly manner later, but at least the other
versions of contentspec are fully parsed now.