Remove DTD support.
Upon closer reading of the spec, it has become clear that it is
essentially impossible to correctly support the DTD section with a
parser that returns slices of the original input (which is a core design
goal of dxml). The recursive replacement of references with the text
that they represent completely kills that, and unlike what I thought
initially, that is required in order to deal with the DTD correctly. We
_could_ alter the parser to recognize where the references could be
(though that would be a royal pain given that the grammar doesn't
actually include that - insanely, the grammar is only the grammar
_after_ all of the references have been replaced) and then just alter
the output so that it could indicate that there were references there
instead of the actual content, but that would get disgusting with stuff
like AttDefInfo. It was reasonable when the references were just
embedded in text and could be replaced by helper functions later, but
when you basically need to fully deal with them to provide stuff like
AttDefInfo and EntityDef, that's really not reasonable.

Really, to do it correctly, the parser needs to transform the text as it
goes along, and potentially, it actually needs to be able able to go
fetch documents from elsewhere (including the internet) in order to grab
all of the documents that the references refer to. And that calls for a
parser that either returns a wrapper range or simply puts everything in
strings as is willing to allocate where necessary. And that's not where
I want dxml to go. Doing that would make the common case worse in order
to support a use case that I personally don't think should even exist.
IMHO, even having the DTD in the XML spec was an enormous mistake. The
simple fact that it makes the parser exponentially more complex should
make that clear even without getting into how it ridiculous it is for a
document to tell the application what constitutes a valid document. It's
the application's job to decide what valid input is. Having the input
decide what valid input is makes no sense.

So, for the sake of simplicity, sanity, and efficiency, I'm stripping
out DTD support. _Maybe_, at some point in the future, it can be added
back as a separate parser that parses the DTD section (without trying to
operate on slices) and generates an object that can be used to validate
the data gotten from EntityCursor, but if so, that's way down the line.

For now, we at least get a _way_ cleaner and shorter parser that still
provides the core functionality that matters.

Much as I'm tempted to strip out the processing instruction support
(since I think that it's basically useless for reasons similar to the
DTD being useless), I'm leaving it in. DTD support can be left out with
the excuse that the parser is not a validating parser, but if it doesn't
support PI's, it's not even a parser that supports all well-formed XML
documents. But fortunately, PI's are dead simple, and of course,
simpleXML will still skip them.