Remove config.skipCommentWS and make it always SkipComentWS.yes.
The more I think about it, the less sense this option makes. Far too
much whitespace is stripped out to reproduce the original document as-is
even with SkipCommentWS.no, and I don't see a valid reason to retain
just whitespace between tags. We can add it back in some form or other
if someone complains about the need later.