Remove contentAsText.
Maybe, we'll add it back later, but clearly, implementing is going to be
a bit involved, since there's no generic way to find out how far apart
two points in a range are, and it's not something that's going to
convert over to the DOM parser very well. So, it seems better to just
drop it for now.