Switch to rejecting all non-predefined entity references.
Given that entity references can be XML that then needs to be further
parsed and can thus affect the result beyond simply being different text
in an attribute value or a text entity, it really doesn't make sense to
let them through unprocessed. We either need to process them or reject
them, and since we're definitely not going to process them, we need to
reject them. So now, they all result in XMLParsingExceptions.