Remove opEquals from EntityRange.Entity.
opEquals was removed from EntityRange.Entity. It didn't handle pos
properly, and in retrospect, it really doesn't make sense. When
comparing entities from the same document, the pos property is
sufficient to detect whether two entities refer to the same point in the
document, and it's unclear what it should mean to compare entities from
separate documents. Ultimately, it seems far better to leave that up to
the programmer.