dxml.parser.stax has been renamed to dxml.parser.
When adding DOM support, it became clear that it made more sense to just
treat the StAX parser as the only parser and the DOM as an object
generated from the parser. So, dxml.parser.stax has been renamed. For
dxml 0.2.0, it will be in dxml/parser/package.d, and dxml/parser/stax.d
will be deprecated and publicly import dxml.parser. In 0.3.0,
dxml/parser/package.d will be moved to dxml/parser.d and
dxml/parsr/stax.d will be removed. That way, we have a single release
with a deprecation warning for those whe grabbed the first release.