Switch to custom unit test build.
For some reason, as evidenced by
https://github.com/jmdavis/dxml/issues/2, projects that depend on dxml
that then do their own unit test builds get linker errors due to stuff
in the unit tests. Those tests really shouldn't even be being built when
dxml is used as a dependency anyway, but for whatever reason, dub builds
all dependencies with the unittest target when a project is built with
the unittest target.

To fix that, dxml now has a separate "doTests" build type which defines
the version dxmlTests, and all unittest blocks are versioned with
dxmlTests. So, unfortunately, dub test will not run the tests anymore,
but we're stuck it seems.