More doc fixes and change the behavior of an empty baseIndent.
Having an empty baseIndent mean no newlines makes it impossible to
insert newlines with an empty indent. Removing the special behavior for
an empty baseIndent fixes that. It does have the downside that there is
then no way to globally turn off newlines for an XMLWriter (Newline.no
must be passed to every function that might insert a newline), but I
think that this way is ultimately better. If it turns out that we really
need a way to turn of newlines for an entire XMLWriter, then I think
that we need to find a way to do that that doesn't disallow having
newlines with no indentantion.