Fix @safety bug in std.util.asDecodedXML.
Previously, the range result type for asDecodedXML had a member variable
which was a slice of another member variable. It was smart enough to
redo the slice in the postblit was called in order to fix the slice when
the struct was copied. However, it failed to take into account the fact
that the struct could be moved, and in that case, the slice would become
garbage, and there is not currently a way to detect the move and fix-up
the slice in that case. So, that range has been changed to use indices
for the static array instead of a dynamic array which slices it.