Improve tests for dxml.writer and fix various bugs.
Most notably:

1. Some code that handled ranges rather than strings didn't actually
work properly with non-strings.

2. writeTaggedText could partially write its results if the tag name was
valid but the text wasn't.

3. We disallowed entity references other than the five standard ones,
which arguably makes sense given that we can't validate any of them, and
none of them are valid if there is no DTD section being written, but in
principle, someone could produce the DTD section themselves (e.g. if it
were copied from somewhere) and write the rest of the XML with dxml, in
which case, it would be perfectly valid for other entity references to
exist. So, we allow them now. I suppose that risks bugs in the case
where someone unknowingly writes an entity tag without realizing that
they need a DTD for it, but if they then parse their own document with
dxml, they'll then know unless they use ThrowOnEntityRef.no.

4. The signatures for several functions were wrong. The template
constraint allowed other range types for tag names even though the
functions didn't - probably left over from before name was required to
be a string.

5. opAssign is now disabled on XMLWriter.