Add public aliases for Attribute type.
EntityRange.Entity now has an alias for the exact Tuple instantiation
that its attributes function returns a range of. DOMEntity also now has
a public Attribute alias.

This will make it easier to refer to the attributes by type if necessary
(e.g. when creating an output range to contain some of them).