Clean up cruft in XMLParsingException.
The constructor still contained some code that was from when dxml had
PositionType, and TextPos could end up with line or column being -1 if
only the line being tracked (and not the column) or if neither was being
tracked. But since that was dropped, the extra logic was pointless, and
it has now been removed.