Remove ldc from .travis.yml due to linker errors.
Even though dxml does work with ldc, the test builds do not. I don't
know if it's some sort of bad interaction between ldc and dub or what,
but until I figure it out, it doesn't make sense to enable it in the
travis stuff anywhere other than master (since having it enabled on the
latest release will cause problems for the official D CI stuff, since
dxml is being added to the list of projects to be tested with that).